refactor: extract a11y audit into dedicated module
Split dom-queries.ts into two focused modules:
- dom-queries.ts: structured DOM querying and page info extraction
- a11y-audit.ts: axe-core loading, accessibility auditing, result formatting

Re-exports from dom-queries.ts preserved for backward compatibility.
